pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Lxg1c <teran_2005@mail.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["all"], version = "^0.115.6"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
aiosqlite = "^0.20.0"
pydantic-settings = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
